refactor(errors): split files to maintain compat with go <1.13
