+++ v2/errors/error_test.go
@@ -271,28 +271,6 @@ func TestUnwrapCustomCause(t *testing.T) {
 	}
 }
 
-func TestFindingErrorInChain(t *testing.T) {
-	baseErr := errors.New("base error")
-	wrappedErr := errors.Wrap(baseErr, "failed")
-	err := New(wrappedErr, 0)
-
-	if !errors.Is(err, baseErr) {
-		t.Errorf("Failed to find base error: %s", err.Error())
-	}
-}
-
-func TestErrorUnwrapping(t *testing.T) {
-	baseErr := errors.New("base error")
-	wrappedErr := fmt.Errorf("failed: %w", baseErr)
-	err := New(wrappedErr, 0)
-
-	unwrapped := errors.Unwrap(err)
-
-	if unwrapped != baseErr {
-		t.Errorf("Failed to find base error: %s", unwrapped.Error())
-	}
-}
-
 func ExampleErrorf() {
 	for i := 1; i <= 2; i++ {
 		if i%2 == 1 {
